+++ ui/ztncui/src/routes/index.js
@@ -33,13 +33,13 @@ router.get('/logout', function(req, res) {
 router.get('/login', guest_only, function(req, res) {
   let message = null;
   if (req.session.error) {
-    if (req.session.error !== 'Access denied!') {
+    if (req.session.error !== '无权访问！') {
       message = req.session.error;
     }
   } else {
     message = req.session.success;
   }
-  res.render('login', { title: 'Login', message: message });
+  res.render('login', { title: '登录', message: message });
 });
 
 router.post('/login', async function(req, res) {
@@ -47,7 +47,7 @@ router.post('/login', async function(req, res) {
     if (user) {
       req.session.regenerate(function() {
         req.session.user = user;
-        req.session.success = 'Authenticated as ' + user.name;
+        req.session.success = '已登录为：' + user.name;
         if (user.pass_set) {
           res.redirect(req.query.redirect || '/controller');
         } else {
@@ -55,7 +55,7 @@ router.post('/login', async function(req, res) {
         }
       });
     } else {
-      req.session.error = 'Authentication failed, please check your username and password.'
+      req.session.error = '登录失败，请检查用户名和密码。'
       res.redirect('/login');
     }
   });